+++ web3/personal.py
@@ -3,7 +3,7 @@ from __future__ import absolute_import
 import getpass
 import warnings
 
-from web3.utils.module import (
+from web3.module import (
     Module,
 )
