+++ web3/txpool.py
@@ -1,4 +1,4 @@
-from web3.utils.module import (
+from web3.module import (
     Module,
 )
