+++ web3/version.py
@@ -1,6 +1,6 @@
 from __future__ import absolute_import
 
-from web3.utils.module import (
+from web3.module import (
     Module,
 )
